Fixed various drawing issues;

+++ app/src/main/java/com/dpanayotov/wallpaper/Circle.java
@@ -4,11 +4,11 @@ package com.dpanayotov.wallpaper;
  * Created by Dean Panayotov Local on 2.9.2015
  */
 public class Circle {
-    public int x;
-    public int y;
+    public float x;
+    public float y;
     public int c;
 
-    public Circle(int x, int y, int c) {
+    public Circle(float x, float y, int c) {
         this.x = x;
         this.y = y;
         this.c = c;
